Refactor CRM, API, and dashboard monoliths
