+++ apps/admin_dashboard/src/main.tsx
@@ -22,7 +22,8 @@ import {
 } from "lucide-react"
 import { type ReactNode, StrictMode, useEffect, useMemo, useRef, useState } from "react"
 import { createRoot } from "react-dom/client"
-
+import { Empty } from "@/components/empty"
+import { SortableTableHead } from "@/components/sortable-table-head"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
@@ -50,6 +51,19 @@ import {
   toneForOnboardingState,
 } from "@/dashboard-utils"
 import { cn } from "@/lib/utils"
+import {
+  type ConfigurationItem,
+  type ConfigurationResponse,
+  ConfigurationView,
+} from "@/views/configuration-view"
+import {
+  type NewsletterStatus,
+  type NewsletterSuppression,
+  type NewsletterSyncPreview,
+  newsletterPreviewSummary,
+  newsletterProviderResults,
+} from "@/views/newsletter-models"
+import { NewsletterView } from "@/views/newsletter-view"
 import "./index.css"
 
 type View =
@@ -72,105 +86,6 @@ type User = {
   crm_contact_id?: string
   crm_base_url?: string
   permissions?: string[]
-}
-
-type ConfigurationItem = {
-  key: string
-  label: string
-  category: string
-  description: string
-  value_type: "string" | "bool" | "int" | "float" | "url" | "csv"
-  is_secret: boolean
-  env_locked: boolean
-  source: "env" | "database" | "default"
-  configured: boolean
-  restart_required: boolean
-  secret_encryption_configured?: boolean | null
-  value?: string | number | boolean | null
-  masked_value?: string | null
-}
-
-type ConfigurationResponse = {
-  items: ConfigurationItem[]
-}
-
-type ConfigurationGroupMetadata = {
-  category: string
-  label: string
-  description: string
-}
-
-const configurationGroups: ConfigurationGroupMetadata[] = [
-  {
-    category: "CRM",
-    label: "CRM",
-    description: "EspoCRM connection settings used by the API, worker, and Discord bot.",
-  },
-  {
-    category: "Projects",
-    label: "Projects",
-    description: "ERPNext credentials and project workflow settings.",
-  },
-  {
-    category: "Onboarding",
-    label: "Onboarding",
-    description:
-      "Editable onboarding integrations such as DocuSeal, Outline, and onboarding email SMTP.",
-  },
-  {
-    category: "Newsletter",
-    label: "Newsletter",
-    description: "Brevo, Keila, and recurring 508 members audience sync settings.",
-  },
-  {
-    category: "AI",
-    label: "AI Providers",
-    description: "Provider credentials, base URLs, and model defaults.",
-  },
-  {
-    category: "Agent",
-    label: "Agent Runtime",
-    description: "Planner, fallback, and tiered model routing for agent workflows.",
-  },
-  {
-    category: "Observability",
-    label: "Observability",
-    description: "Telemetry and request tracing integrations.",
-  },
-  {
-    category: "Intake",
-    label: "Intake",
-    description: "Resume and mailbox intake limits and parser defaults.",
-  },
-  {
-    category: "Operations",
-    label: "Operations",
-    description: "Queue, sync, GitHub, and notification behavior.",
-  },
-  {
-    category: "Legacy",
-    label: "Legacy",
-    description: "Older integrations retained for compatibility.",
-  },
-]
-
-const configurationGroupByCategory = new Map(
-  configurationGroups.map((group, index) => [group.category, { ...group, index }]),
-)
-
-function configurationGroupId(category: string) {
-  return `configurationGroup-${category.replace(/[^a-zA-Z0-9_-]+/g, "-")}`
-}
-
-function isPrimaryConfiguration(item: ConfigurationItem) {
-  return (
-    item.key.startsWith("ONBOARDING_EMAIL_") ||
-    item.is_secret ||
-    item.value_type === "url" ||
-    item.key.endsWith("_MODEL") ||
-    item.key.endsWith("_API_USER") ||
-    item.key.endsWith("_BASE_URL")
-  )
 }
 
 type ProfileStatus = {
@@ -441,47 +356,6 @@ type WikiMatchPreview = {
   }>
 }
 
-type NewsletterSyncPreview = {
-  mailboxes_scanned?: number
-  contacts_considered?: number
-  providers?: Record<
-    string,
-    {
-      would_sync?: number
-      synced?: number
-      skipped?: number
-      failed?: number
-    }
-  >
-}
-
-type NewsletterSuppression = {
-  email: string
-  source_provider: string
-  reason: string
-  active?: boolean
-  first_seen_at?: string
-  last_seen_at?: string
-  updated_at?: string
-  metadata?: Record<string, unknown>
-}
-
-type NewsletterStatus = {
-  scheduler_enabled?: boolean
-  interval_seconds?: number
-  active_suppression_count?: number
-  active_suppressed_email_count?: number
-  latest_job?: JobDetail | null
-}
-
-type NewsletterProviderResult = {
-  synced?: number
-  would_sync?: number
-  skipped?: number
-  failed?: number
-  statuses?: Record<string, number>
-}
-
 type AuditEvent = {
   id?: string
   occurred_at?: string
@@ -702,52 +576,6 @@ function messageFromUnknown(error: unknown, fallback: string) {
   return fallback
 }
 
-function newsletterPreviewSummary(preview: NewsletterSyncPreview | undefined) {
-  if (!preview) return null
-  const providerSummaries = Object.entries(preview.providers || {}).map(([name, result]) => {
-    const wouldSync = result.would_sync ?? result.synced ?? 0
-    return `${name}: ${wouldSync} would sync, ${result.skipped || 0} skipped, ${result.failed || 0} failed`
-  })
-  const scanned = preview.mailboxes_scanned ?? 0
-  const contacts = preview.contacts_considered ?? 0
-  return [
-    `${scanned} mailbox${scanned === 1 ? "" : "es"}`,
-    `${contacts} contact${contacts === 1 ? "" : "s"}`,
-    ...providerSummaries,
-  ].join("; ")
-}
-
-function newsletterIntervalLabel(seconds?: number) {
-  const normalized = Number(seconds || 0)
-  if (!Number.isFinite(normalized) || normalized <= 0) return "Not configured"
-  if (normalized % 86400 === 0) {
-    const days = normalized / 86400
-    return `${days} day${days === 1 ? "" : "s"}`
-  }
-  if (normalized % 3600 === 0) {
-    const hours = normalized / 3600
-    return `${hours} hour${hours === 1 ? "" : "s"}`
-  }
-  if (normalized % 60 === 0) {
-    const minutes = normalized / 60
-    return `${minutes} minute${minutes === 1 ? "" : "s"}`
-  }
-  return `${normalized} second${normalized === 1 ? "" : "s"}`
-}
-
-function newsletterProviderResults(job?: JobDetail | null) {
-  const result = job?.result
-  if (!result || typeof result !== "object" || Array.isArray(result)) return []
-  const providers = (result as { providers?: unknown }).providers
-  if (!providers || typeof providers !== "object" || Array.isArray(providers)) return []
-  return Object.entries(providers)
-    .filter((entry): entry is [string, NewsletterProviderResult] => {
-      const [, value] = entry
-      return Boolean(value && typeof value === "object" && !Array.isArray(value))
-    })
-    .sort(([left], [right]) => left.localeCompare(right))
-}
-
 function devErrorFromUnknown(error: unknown, fallback: string): DashboardDevError {
   const message = messageFromUnknown(error, fallback)
   const apiError = error instanceof ApiRequestError ? error : null
@@ -929,58 +757,6 @@ function sortItems<T extends SortableItem>(
   })
 }
 
-function SortButton({
-  label,
-  scope,
-  sort,
-  sortKey,
-  onSort,
-}: {
-  label: string
-  scope: View
-  sort: { key: string; direction: SortDirection }
-  sortKey: string
-  onSort: (scope: View, key: string) => void
-}) {
-  const active = sort.key === sortKey
-  const arrow = sort.direction === "asc" ? "↑" : "↓"
-  return (
-    <button
-      type="button"
-      data-sort-scope={scope}
-      data-sort-key={sortKey}
-      className="text-left font-[inherit] text-inherit hover:text-foreground"
-      onClick={() => onSort(scope, sortKey)}
-    >
-      {active ? `${label} ${arrow}` : label}
-    </button>
-  )
-}
-
-function SortableTableHead({
-  className,
-  label,
-  scope,
-  sort,
-  sortKey,
-  onSort,
-}: {
-  className?: string
-  label: string
-  scope: View
-  sort: { key: string; direction: SortDirection }
-  sortKey: string
-  onSort: (scope: View, key: string) => void
-}) {
-  const ariaSort =
-    sort.key === sortKey ? (sort.direction === "asc" ? "ascending" : "descending") : "none"
-  return (
-    <TableHead className={className} aria-sort={ariaSort}>
-      <SortButton label={label} scope={scope} sort={sort} sortKey={sortKey} onSort={onSort} />
-    </TableHead>
-  )
-}
-
 function Metric({ label, value, id }: { label: string; value: number; id?: string }) {
   return (
     <Card className="p-4">
@@ -990,11 +766,6 @@ function Metric({ label, value, id }: { label: string; value: number; id?: strin
       </strong>
     </Card>
   )
-}
-
-function Empty({ children, hidden }: { children: string; hidden: boolean }) {
-  if (hidden) return null
-  return <div className="px-4 py-7 text-center text-sm text-muted-foreground">{children}</div>
 }
 
 function HighlightedText({ value, query }: { value: string; query: string }) {
@@ -5982,283 +5753,6 @@ function PeopleView(props: {
   )
 }
 
-function NewsletterView(props: {
-  status: NewsletterStatus | null
-  suppressions: NewsletterSuppression[]
-  providerOptions: string[]
-  providerFilter: string
-  sort: { key: string; direction: SortDirection }
-  loading: Record<string, boolean>
-  canSync: boolean
-  onRefresh: () => void
-  onSync: () => void
-  onProviderFilterChange: (value: string) => void
-  onSort: (key: string) => void
-}) {
-  const latestJob = props.status?.latest_job || null
-  const uniqueSuppressed = props.status?.active_suppressed_email_count
-  const suppressionRows = props.status?.active_suppression_count
-  const providerResults = newsletterProviderResults(latestJob)
-  const providerOptions = props.providerFilter
-    ? [...new Set([...props.providerOptions, props.providerFilter])].sort((left, right) =>
-        left.localeCompare(right),
-      )
-    : props.providerOptions
-  const suppressionsByProvider = props.suppressions.reduce<Record<string, NewsletterSuppression[]>>(
-    (groups, record) => {
-      const provider = record.source_provider || "unknown"
-      groups[provider] = [...(groups[provider] || []), record]
-      return groups
-    },
-    {},
-  )
-  const suppressionGroups = Object.entries(suppressionsByProvider).sort(([left], [right]) =>
-    left.localeCompare(right),
-  )
-  return (
-    <>
-      <Card>
-        <CardHeader>
-          <CardTitle>Newsletter sync</CardTitle>
-          <div className="flex flex-wrap items-center justify-end gap-2">
-            <Button
-              id="refreshNewsletter"
-              type="button"
-              variant="outline"
-              onClick={props.onRefresh}
-              disabled={props.loading.newsletterStatus || props.loading.newsletterSuppressions}
-            >
-              <RefreshCw />
-              Refresh
-            </Button>
-            {props.canSync ? (
-              <Button
-                id="syncNewsletters"
-                data-permission="people:sync"
-                type="button"
-                onClick={props.onSync}
-                disabled={props.loading.syncNewsletters}
-              >
-                <RefreshCw />
-                Sync now
-              </Button>
-            ) : null}
-          </div>
-        </CardHeader>
-        <CardContent>
-          <div className="grid gap-3 md:grid-cols-4">
-            <div className="rounded-md border bg-background p-4">
-              <span className="text-xs font-bold text-muted-foreground">Scheduler</span>
-              <strong id="newsletterScheduler" className="block text-2xl">
-                {props.status?.scheduler_enabled ? "Enabled" : "Disabled"}
-              </strong>
-            </div>
-            <div className="rounded-md border bg-background p-4">
-              <span className="text-xs font-bold text-muted-foreground">Interval</span>
-              <strong id="newsletterInterval" className="block text-2xl">
-                {newsletterIntervalLabel(props.status?.interval_seconds)}
-              </strong>
-            </div>
-            <div className="rounded-md border bg-background p-4">
-              <span className="text-xs font-bold text-muted-foreground">Suppressed emails</span>
-              <strong id="newsletterSuppressedEmails" className="block text-2xl">
-                {uniqueSuppressed ?? "Loading"}
-              </strong>
-            </div>
-            <div className="rounded-md border bg-background p-4">
-              <span className="text-xs font-bold text-muted-foreground">Suppression rows</span>
-              <strong id="newsletterSuppressionRows" className="block text-2xl">
-                {suppressionRows ?? "Loading"}
-              </strong>
-            </div>
-          </div>
-          <div className="mt-4 grid gap-2 rounded-md border bg-secondary p-3 text-sm">
-            <div className="flex flex-wrap items-center gap-2">
-              <span className="font-extrabold">Latest sync</span>
-              {latestJob ? (
-                <Badge variant={(latestJob.status as Tone) || "neutral"}>{latestJob.status}</Badge>
-              ) : (
-                <Badge variant="neutral">No job found</Badge>
-              )}
-            </div>
-            {latestJob ? (
-              <>
-                <div className="text-muted-foreground">
-                  {latestJob.type} | updated {formatDate(latestJob.updated_at)} | attempts{" "}
-                  {latestJob.attempts}/{latestJob.max_attempts}
-                </div>
-                {latestJob.last_error ? (
-                  <div className="text-red-700 dark:text-red-300">{latestJob.last_error}</div>
-                ) : null}
-                <div className="grid gap-2 md:grid-cols-2">
-                  {providerResults.length ? (
-                    providerResults.map(([providerName, providerResult]) => {
-                      const statuses = providerResult.statuses || {}
-                      const statusEntries = Object.entries(statuses).sort(([left], [right]) =>
-                        left.localeCompare(right),
-                      )
-                      const synced = providerResult.synced ?? providerResult.would_sync ?? 0
-                      return (
-                        <div
-                          key={providerName}
-                          className="grid gap-2 rounded-md border bg-background p-3"
-                        >
-                          <div className="flex flex-wrap items-center justify-between gap-2">
-                            <strong>{providerName}</strong>
-                            <div className="flex flex-wrap gap-1">
-                              <Badge variant="succeeded">{synced} synced</Badge>
-                              <Badge variant="neutral">{providerResult.skipped || 0} skipped</Badge>
-                              {providerResult.failed ? (
-                                <Badge variant="failed">{providerResult.failed} failed</Badge>
-                              ) : null}
-                            </div>
-                          </div>
-                          {statusEntries.length ? (
-                            <div className="flex flex-wrap gap-1">
-                              {statusEntries.map(([status, count]) => (
-                                <Badge key={status} variant="neutral">
-                                  {status.replaceAll("_", " ")}: {count}
-                                </Badge>
-                              ))}
-                            </div>
-                          ) : null}
-                        </div>
-                      )
-                    })
-                  ) : (
-                    <div className="rounded-md border bg-background p-3 text-muted-foreground">
-                      No provider status details recorded for the latest sync.
-                    </div>
-                  )}
-                </div>
-                {latestJob.result ? (
-                  <pre className="max-h-40 overflow-auto rounded-md bg-background p-3 text-xs">
-                    {jsonPreview(latestJob.result)}
-                  </pre>
-                ) : null}
-              </>
-            ) : (
-              <div className="text-muted-foreground">
-                No 508 members newsletter sync job was found in the last 90 days.
-              </div>
-            )}
-          </div>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader>
-          <CardTitle>Newsletter suppressions</CardTitle>
-          <div className="flex flex-wrap items-center justify-end gap-3">
-            <Label className="min-w-44">
-              Provider
-              <Select
-                id="newsletterProviderFilter"
-                value={props.providerFilter}
-                onChange={(event) => props.onProviderFilterChange(event.target.value)}
-              >
-                <option value="">All providers</option>
-                {providerOptions.map((provider) => (
-                  <option key={provider} value={provider}>
-                    {provider}
-                  </option>
-                ))}
-              </Select>
-            </Label>
-            <span id="newsletterSuppressionsStatus" className="text-sm text-muted-foreground">
-              {props.loading.newsletterSuppressions
-                ? "Loading"
-                : `${props.suppressions.length} shown`}
-            </span>
-          </div>
-        </CardHeader>
-        <Empty hidden={props.suppressions.length !== 0}>
-          No active newsletter suppressions recorded.
-        </Empty>
-        <div className={cn("grid gap-4", props.suppressions.length === 0 && "hidden")}>
-          {suppressionGroups.map(([provider, records]) => (
-            <section key={provider} className="grid gap-2" aria-label={`${provider} suppressions`}>
-              <div className="flex flex-wrap items-center gap-2 px-4 pt-4">
-                <Badge variant="neutral">{provider}</Badge>
-                <span className="text-sm text-muted-foreground">
-                  {records.length} suppression{records.length === 1 ? "" : "s"}
-                </span>
-              </div>
-              <div className="overflow-x-auto">
-                <Table
-                  id={`newsletterSuppressionsTable-${provider}`}
-                  className="min-w-[900px]"
-                  aria-label={`${provider} newsletter suppressions`}
-                >
-                  <TableHeader>
-                    <TableRow>
-                      <SortableTableHead
-                        className="w-[34%]"
-                        label="Email"
-                        scope="newsletter"
-                        sort={props.sort}
-                        sortKey="email"
-                        onSort={(_, key) => props.onSort(key)}
-                      />
-                      <SortableTableHead
-                        className="w-[14%]"
-                        label="Source"
-                        scope="newsletter"
-                        sort={props.sort}
-                        sortKey="source_provider"
-                        onSort={(_, key) => props.onSort(key)}
-                      />
-                      <SortableTableHead
-                        className="w-[22%]"
-                        label="Reason"
-                        scope="newsletter"
-                        sort={props.sort}
-                        sortKey="reason"
-                        onSort={(_, key) => props.onSort(key)}
-                      />
-                      <SortableTableHead
-                        className="w-[15%]"
-                        label="First seen"
-                        scope="newsletter"
-                        sort={props.sort}
-                        sortKey="first_seen_at"
-                        onSort={(_, key) => props.onSort(key)}
-                      />
-                      <SortableTableHead
-                        className="w-[15%]"
-                        label="Last seen"
-                        scope="newsletter"
-                        sort={props.sort}
-                        sortKey="last_seen_at"
-                        onSort={(_, key) => props.onSort(key)}
-                      />
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody id={`newsletterSuppressionsBody-${provider}`}>
-                    {records.map((record) => (
-                      <TableRow key={`${record.email}-${record.source_provider}`}>
-                        <TableCell className="font-mono text-sm">{record.email}</TableCell>
-                        <TableCell>
-                          <Badge variant="neutral">{record.source_provider}</Badge>
-                        </TableCell>
-                        <TableCell>{record.reason.replaceAll("_", " ")}</TableCell>
-                        <TableCell>{formatDate(record.first_seen_at)}</TableCell>
-                        <TableCell>
-                          {formatDate(record.last_seen_at || record.updated_at)}
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </div>
-            </section>
-          ))}
-        </div>
-      </Card>
-    </>
-  )
-}
-
 function OnboardingView(props: {
   people: Person[]
   sort: { key: string; direction: SortDirection }
@@ -7723,471 +7217,6 @@ function AgentView({
   )
 }
 
-function ConfigurationView({
-  items,
-  loading,
-  canWrite,
-  onRefresh,
-  onSave,
-  onClear,
-  focusCategory,
-  focusNonce,
-}: {
-  items: ConfigurationItem[]
-  loading: Record<string, boolean>
-  canWrite: boolean
-  onRefresh: () => void
-  onSave: (key: string, value: string) => Promise<boolean>
-  onClear: (key: string) => void
-  focusCategory?: string
-  focusNonce?: number
-}) {
-  const [selectedCategory, setSelectedCategory] = useState("All")
-  const [highlightedCategory, setHighlightedCategory] = useState("")
-  const [drafts, setDrafts] = useState<Record<string, string>>({})
-  const [generatedSecrets, setGeneratedSecrets] = useState<Record<string, string>>({})
-  const categories = useMemo(() => {
-    const present = new Set(items.map((item) => item.category))
-    const known = configurationGroups.filter((group) => present.has(group.category))
-    const unknown = Array.from(present)
-      .filter((category) => !configurationGroupByCategory.has(category))
-      .sort()
-      .map((category) => ({
-        category,
-        label: category,
-        description: "Additional runtime settings.",
-      }))
-    return known.concat(unknown)
-  }, [items])
-  const groupedItems = useMemo(() => {
-    const groups = new Map<string, ConfigurationItem[]>()
-    for (const item of items) {
-      if (selectedCategory !== "All" && item.category !== selectedCategory) continue
-      const current = groups.get(item.category) ?? []
-      current.push(item)
-      groups.set(item.category, current)
-    }
-    return Array.from(groups.entries())
-      .map(([category, groupItems]) => {
-        const known = configurationGroupByCategory.get(category)
-        const sortedItems = groupItems.sort((left, right) => {
-          const priorityOrder =
-            Number(!isPrimaryConfiguration(left)) - Number(!isPrimaryConfiguration(right))
-          return priorityOrder || left.label.localeCompare(right.label)
-        })
-        const primaryItems = sortedItems.filter(isPrimaryConfiguration)
-        const advancedItems = sortedItems.filter((item) => !isPrimaryConfiguration(item))
-        return {
-          category,
-          label: known?.label ?? category,
-          description: known?.description ?? "Additional runtime settings.",
-          order: known?.index ?? configurationGroups.length,
-          primaryItems: primaryItems.length ? primaryItems : sortedItems,
-          advancedItems: primaryItems.length ? advancedItems : [],
-          items: sortedItems,
-        }
-      })
-      .sort((left, right) => left.order - right.order || left.label.localeCompare(right.label))
-  }, [items, selectedCategory])
-  const summary = useMemo(
-    () => ({
-      configured: items.filter((item) => item.configured).length,
-      envLocked: items.filter((item) => item.env_locked).length,
-      missing: items.filter((item) => !item.configured).length,
-    }),
-    [items],
-  )
-  const visibleItemCount = groupedItems.reduce((count, group) => count + group.items.length, 0)
-
-  useEffect(() => {
-    setDrafts(
-      Object.fromEntries(
-        items.map((item) => [item.key, item.is_secret ? "" : String(item.value ?? "")]),
-      ),
-    )
-  }, [items])
-
-  useEffect(() => {
-    if (selectedCategory !== "All" && !items.some((item) => item.category === selectedCategory)) {
-      setSelectedCategory("All")
-    }
-  }, [items, selectedCategory])
-
-  useEffect(() => {
-    if (!focusCategory || !items.some((item) => item.category === focusCategory)) return
-    void focusNonce
-    setSelectedCategory(focusCategory)
-    setHighlightedCategory(focusCategory)
-    const frame = window.requestAnimationFrame?.(() => {
-      const element = document.getElementById(configurationGroupId(focusCategory))
-      if (typeof element?.scrollIntoView === "function") {
-        element.scrollIntoView({
-          block: "start",
-          behavior: "smooth",
-        })
-      }
-    })
-    const timeout = window.setTimeout(() => setHighlightedCategory(""), 4000)
-    return () => {
-      if (frame !== undefined) window.cancelAnimationFrame?.(frame)
-      window.clearTimeout(timeout)
-    }
-  }, [focusCategory, focusNonce, items])
-
-  function sourceBadge(item: ConfigurationItem) {
-    if (item.source === "env") return "ENV"
-    if (item.source === "database") return "DB"
-    return "Default"
-  }
-
-  function generateSigningSecret() {
-    const bytes = new Uint8Array(32)
-    window.crypto.getRandomValues(bytes)
-    return Array.from(bytes, (byte) => byte.toString(16).padStart(2, "0")).join("")
-  }
-
-  async function generateTallySigningSecret(item: ConfigurationItem) {
-    const secret = generateSigningSecret()
-    setDrafts((current) => ({ ...current, [item.key]: secret }))
-    const saved = await onSave(item.key, secret)
-    if (saved) {
-      setGeneratedSecrets((current) => ({ ...current, [item.key]: secret }))
-    }
-  }
-
-  async function copyGeneratedSecret(key: string) {
-    const secret = generatedSecrets[key]
-    if (!secret) return
-    if (navigator.clipboard?.writeText) {
-      try {
-        await navigator.clipboard.writeText(secret)
-        return
-      } catch {
-        // Fall through to selecting the visible field.
-      }
-    }
-    const element = document.getElementById(`generatedSecret-${key}`) as HTMLInputElement | null
-    element?.select()
-  }
-
-  function valueInput(item: ConfigurationItem) {
-    const value = drafts[item.key] ?? ""
-    const disabled = !canWrite || item.env_locked || loading[`configuration:${item.key}`]
-    if (item.value_type === "bool") {
-      return (
-        <Select
-          aria-label={`${item.label} value`}
-          value={value}
-          disabled={disabled}
-          onChange={(event) =>
-            setDrafts((current) => ({ ...current, [item.key]: event.target.value }))
-          }
-        >
-          <option value="">Default</option>
-          <option value="true">True</option>
-          <option value="false">False</option>
-        </Select>
-      )
-    }
-    return (
-      <Input
-        aria-label={`${item.label} value`}
-        value={value}
-        type={item.is_secret ? "password" : item.value_type === "int" ? "number" : "text"}
-        inputMode={item.value_type === "int" || item.value_type === "float" ? "numeric" : "text"}
-        placeholder={item.is_secret ? "Set new value" : ""}
-        autoComplete="off"
-        disabled={disabled}
-        onChange={(event) =>
-          setDrafts((current) => ({ ...current, [item.key]: event.target.value }))
-        }
-      />
-    )
-  }
-
-  function configurationTable(groupLabel: string, tableItems: ConfigurationItem[], suffix: string) {
-    return (
-      <div className="overflow-x-auto">
-        <Table
-          id={`configurationTable-${suffix}`}
-          className="min-w-[980px]"
-          aria-label={`${groupLabel} configuration settings`}
-        >
-          <TableHeader>
-            <TableRow>
-              <TableHead className="w-[26%]">Setting</TableHead>
-              <TableHead className="w-[12%]">Source</TableHead>
-              <TableHead className="w-[18%]">Active</TableHead>
-              <TableHead className="w-[29%]">Value</TableHead>
-              <TableHead className="w-[15%]">Actions</TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody id={`configurationBody-${suffix}`}>
-            {tableItems.map((item) => renderConfigurationRow(item))}
-          </TableBody>
-        </Table>
-      </div>
-    )
-  }
-
-  function renderConfigurationRow(item: ConfigurationItem) {
-    const busy = loading[`configuration:${item.key}`]
-    const writable = canWrite && !item.env_locked && !busy
-    const draft = drafts[item.key] ?? ""
-    const emptyNonSecretDraft = !item.is_secret && !draft.trim()
-    const generatedSecret = generatedSecrets[item.key] || ""
-    const showTallySecretGenerator = item.key === "ONBOARDING_TALLY_WEBHOOK_SIGNING_SECRET"
-    return (
-      <TableRow key={item.key}>
-        <TableCell>
-          <div className="grid gap-1">
-            <strong>{item.label}</strong>
-            <span className="font-mono text-xs text-muted-foreground">{item.key}</span>
-            <span className="text-xs text-muted-foreground">{item.description}</span>
-            {item.restart_required ? (
-              <div>
-                <Badge variant="running">Restart</Badge>
-              </div>
-            ) : null}
-          </div>
-        </TableCell>
-        <TableCell>
-          <div className="grid gap-1.5">
-            <Badge variant={item.source === "env" ? "running" : "neutral"}>
-              {sourceBadge(item)}
-            </Badge>
-            {item.env_locked ? (
-              <span className="text-xs text-muted-foreground">Environment locked</span>
-            ) : null}
-          </div>
-        </TableCell>
-        <TableCell>
-          <div className="grid gap-1">
-            <Badge variant={item.configured ? "succeeded" : "missing"}>
-              {item.configured ? "Configured" : "Missing"}
-            </Badge>
-            {item.is_secret ? (
-              <span className="font-mono text-xs text-muted-foreground">
-                {item.masked_value || (item.configured ? "Hidden" : "No secret")}
-              </span>
-            ) : (
-              <span className="break-words text-xs text-muted-foreground">
-                {String(item.value ?? "") || "Default"}
-              </span>
-            )}
-            {item.is_secret && item.secret_encryption_configured === false ? (
-              <span className="text-xs text-red-300">Encryption key missing</span>
-            ) : null}
-          </div>
-        </TableCell>
-        <TableCell>
-          <div className="grid gap-2">
-            {valueInput(item)}
-            {showTallySecretGenerator && generatedSecret ? (
-              <div className="grid gap-2 rounded-md border bg-secondary/30 p-2 text-xs">
-                <span className="font-extrabold">Copy this secret into Tally now.</span>
-                <Input
-                  id={`generatedSecret-${item.key}`}
-                  value={generatedSecret}
-                  readOnly
-                  className="font-mono"
-                  aria-label="Generated Tally webhook signing secret"
-                />
-                <span className="text-muted-foreground">
-                  It is only shown until this page refreshes or you dismiss it.
-                </span>
-              </div>
-            ) : null}
-          </div>
-        </TableCell>
-        <TableCell>
-          <div className="flex flex-wrap justify-end gap-2">
-            {showTallySecretGenerator ? (
-              <Button
-                type="button"
-                size="sm"
-                variant="outline"
-                onClick={() => void generateTallySigningSecret(item)}
-                disabled={
-                  !writable ||
-                  item.secret_encryption_configured === false ||
-                  !window.crypto?.getRandomValues
-                }
-              >
-                Generate
-              </Button>
-            ) : null}
-            {showTallySecretGenerator && generatedSecret ? (
-              <>
-                <Button
-                  type="button"
-                  size="sm"
-                  variant="outline"
-                  onClick={() => void copyGeneratedSecret(item.key)}
-                >
-                  Copy
-                </Button>
-                <Button
-                  type="button"
-                  size="sm"
-                  variant="ghost"
-                  onClick={() =>
-                    setGeneratedSecrets((current) => {
-                      const next = { ...current }
-                      delete next[item.key]
-                      return next
-                    })
-                  }
-                >
-                  Hide
-                </Button>
-              </>
-            ) : null}
-            <Button
-              type="button"
-              size="sm"
-              onClick={() => onSave(item.key, draft)}
-              disabled={
-                !writable ||
-                (item.is_secret && !draft.trim()) ||
-                emptyNonSecretDraft ||
-                (item.is_secret && item.secret_encryption_configured === false)
-              }
-            >
-              Save
-            </Button>
-            <Button
-              type="button"
-              size="sm"
-              variant="outline"
-              onClick={() => onClear(item.key)}
-              disabled={!writable || item.source !== "database"}
-            >
-              Clear
-            </Button>
-          </div>
-        </TableCell>
-      </TableRow>
-    )
-  }
-
-  return (
-    <div className="grid gap-4">
-      <Card>
-        <CardHeader>
-          <CardTitle>Configuration</CardTitle>
-          <Button
-            id="refreshConfiguration"
-            type="button"
-            variant="outline"
-            onClick={onRefresh}
-            disabled={loading.configuration}
-          >
-            <RefreshCw />
-            Refresh
-          </Button>
-        </CardHeader>
-        <CardContent className="grid gap-4">
-          <section
-            className="grid gap-3 sm:grid-cols-2 lg:grid-cols-4"
-            aria-label="Configuration summary"
-          >
-            {[
-              ["Total", items.length],
-              ["Configured", summary.configured],
-              ["Missing", summary.missing],
-              ["Env locked", summary.envLocked],
-            ].map(([label, value]) => (
-              <div key={label} className="rounded-md border bg-background p-3">
-                <span className="text-[11px] font-extrabold uppercase text-muted-foreground">
-                  {label}
-                </span>
-                <strong className="mt-1 block text-xl">{value}</strong>
-              </div>
-            ))}
-          </section>
-          <section className="flex flex-wrap gap-2" aria-label="Configuration groups">
-            <Button
-              type="button"
-              size="sm"
-              variant={selectedCategory === "All" ? "default" : "outline"}
-              aria-pressed={selectedCategory === "All"}
-              onClick={() => setSelectedCategory("All")}
-            >
-              All groups
-              <span className="font-mono text-[11px]">{items.length}</span>
-            </Button>
-            {categories.map((group) => {
-              const count = items.filter((item) => item.category === group.category).length
-              return (
-                <Button
-                  key={group.category}
-                  type="button"
-                  size="sm"
-                  variant={selectedCategory === group.category ? "default" : "outline"}
-                  aria-pressed={selectedCategory === group.category}
-                  onClick={() => setSelectedCategory(group.category)}
-                >
-                  {group.label}
-                  <span className="font-mono text-[11px]">{count}</span>
-                </Button>
-              )
-            })}
-          </section>
-        </CardContent>
-      </Card>
-
-      <Empty hidden={visibleItemCount !== 0}>No configuration entries found.</Empty>
-      {groupedItems.map((group) => {
-        const configured = group.items.filter((item) => item.configured).length
-        const missing = group.items.length - configured
-        const restartRequired = group.items.some((item) => item.restart_required)
-        return (
-          <Card
-            key={group.category}
-            id={configurationGroupId(group.category)}
-            className={cn(
-              "scroll-mt-4 transition-shadow",
-              highlightedCategory === group.category &&
-                "ring-2 ring-primary ring-offset-2 ring-offset-background",
-            )}
-          >
-            <CardHeader className="items-start">
-              <div className="grid gap-1">
-                <CardTitle>{group.label}</CardTitle>
-                <span className="text-sm text-muted-foreground">{group.description}</span>
-              </div>
-              <div className="flex flex-wrap justify-end gap-1.5">
-                <Badge variant="neutral">{group.items.length} settings</Badge>
-                <Badge variant={missing ? "missing" : "succeeded"}>{configured} configured</Badge>
-                {restartRequired ? <Badge variant="running">Restart</Badge> : null}
-              </div>
-            </CardHeader>
-            {configurationTable(group.label, group.primaryItems, group.category)}
-            {group.advancedItems.length ? (
-              <details className="border-t bg-background/40">
-                <summary className="flex min-h-11 cursor-pointer items-center justify-between gap-3 px-4 py-3 text-sm font-extrabold">
-                  <span>Advanced</span>
-                  <span className="font-mono text-xs text-muted-foreground">
-                    {group.advancedItems.length} settings
-                  </span>
-                </summary>
-                <div className="border-t">
-                  {configurationTable(
-                    `${group.label} advanced`,
-                    group.advancedItems,
-                    `${group.category}-advanced`,
-                  )}
-                </div>
-              </details>
-            ) : null}
-          </Card>
-        )
-      })}
-    </div>
-  )
-}
-
 const root = document.getElementById("root")
 
 if (!root) {
@@ -8202,5 +7231,5 @@ if (!root) {
   )
 }
 
-export type { ConfigurationItem }
-export { ConfigurationView }
+export type { ConfigurationItem } from "@/views/configuration-view"
+export { ConfigurationView } from "@/views/configuration-view"
